+++ components/tour/index.ts
@@ -1,2 +1,12 @@
 export { Tour } from './Tour'
-export type { TourProps, TourStep, TourButtonProps, TourPlacement, TourClassNames, TourStyles } from './types'
+export type {
+  TourProps,
+  TourStep,
+  TourButtonProps,
+  TourPlacement,
+  TourClassNames,
+  TourStyles,
+  TourChangeHandler,
+  TourCloseHandler,
+  TourFinishHandler,
+} from './types'